+++ android/src/main/java/top/huic/tencent_im_plugin/TencentImPlugin.java
@@ -24,6 +24,7 @@ import com.tencent.imsdk.v2.V2TIMGroupMemberInfoResult;
 import com.tencent.imsdk.v2.V2TIMGroupMemberOperationResult;
 import com.tencent.imsdk.v2.V2TIMManager;
 import com.tencent.imsdk.v2.V2TIMMessage;
+import com.tencent.imsdk.v2.V2TIMMessageListGetOption;
 import com.tencent.imsdk.v2.V2TIMOfflinePushConfig;
 import com.tencent.imsdk.v2.V2TIMOfflinePushInfo;
 import com.tencent.imsdk.v2.V2TIMSDKConfig;
@@ -77,6 +78,36 @@ public class TencentImPlugin implements FlutterPlugin, MethodCallHandler {
      * 与Flutter的通信管道
      */
     private static MethodChannel channel;
+
+    /**
+     * Sdk监听器
+     */
+    private final CustomSDKListener sdkListener = new CustomSDKListener();
+
+    /**
+     * 消息监听器
+     */
+    private final CustomAdvancedMsgListener advancedMsgListener = new CustomAdvancedMsgListener();
+
+    /**
+     * 会话监听器
+     */
+    private final CustomConversationListener conversationListener = new CustomConversationListener();
+
+    /**
+     * 群监听器
+     */
+    private final CustomGroupListener groupListener = new CustomGroupListener();
+
+    /**
+     * 关系链监听器
+     */
+    private final CustomFriendshipListener friendshipListener = new CustomFriendshipListener();
+
+    /**
+     * 信令监听器
+     */
+    private final CustomSignalingListener signalingListener = new CustomSignalingListener();
 
     public TencentImPlugin() {
     }
@@ -141,22 +172,22 @@ public class TencentImPlugin implements FlutterPlugin, MethodCallHandler {
         if (logPrintLevel != null) {
             sdkConfig.setLogLevel(logPrintLevel);
         }
-        V2TIMManager.getInstance().initSDK(context, Integer.parseInt(appid), sdkConfig, new CustomSDKListener());
+        V2TIMManager.getInstance().initSDK(context, Integer.parseInt(appid), sdkConfig, sdkListener);
 
         // 绑定消息监听
-        V2TIMManager.getMessageManager().addAdvancedMsgListener(new CustomAdvancedMsgListener());
+        V2TIMManager.getMessageManager().addAdvancedMsgListener(this.advancedMsgListener);
 
         // 绑定会话监听器
-        V2TIMManager.getConversationManager().setConversationListener(new CustomConversationListener());
+        V2TIMManager.getConversationManager().setConversationListener(this.conversationListener);
 
         // 绑定群监听器
-        V2TIMManager.getInstance().setGroupListener(new CustomGroupListener());
+        V2TIMManager.getInstance().setGroupListener(this.groupListener);
 
         // 绑定关系链监听器
-        V2TIMManager.getFriendshipManager().setFriendListener(new CustomFriendshipListener());
+        V2TIMManager.getFriendshipManager().setFriendListener(this.friendshipListener);
 
         // 绑定信令监听器
-        V2TIMManager.getSignalingManager().addSignalingListener(new CustomSignalingListener());
+        V2TIMManager.getSignalingManager().addSignalingListener(this.signalingListener);
 
         result.success(null);
     }
@@ -170,7 +201,29 @@ public class TencentImPlugin implements FlutterPlugin, MethodCallHandler {
      */
     private void unInitSDK(MethodCall methodCall, Result result) {
         V2TIMManager.getInstance().unInitSDK();
+        V2TIMManager.getMessageManager().removeAdvancedMsgListener(this.advancedMsgListener);
+        V2TIMManager.getSignalingManager().removeSignalingListener(this.signalingListener);
         result.success(null);
+    }
+
+    /**
+     * 获得SDK版本
+     *
+     * @param methodCall 方法调用对象
+     * @param result     返回结果对象
+     */
+    private void getVersion(MethodCall methodCall, Result result) {
+        result.success(V2TIMManager.getInstance().getVersion());
+    }
+
+    /**
+     * 获得服务器当前时间
+     *
+     * @param methodCall 方法调用对象
+     * @param result     返回结果对象
+     */
+    private void getServerTime(MethodCall methodCall, Result result) {
+        result.success(V2TIMManager.getInstance().getServerTime());
     }
 
     /**
@@ -249,7 +302,7 @@ public class TencentImPlugin implements FlutterPlugin, MethodCallHandler {
      */
     private void inviteInGroup(MethodCall methodCall, final Result result) {
         String groupID = CommonUtil.getParam(methodCall, result, "groupID");
-        List<String> inviteeList = Arrays.asList(CommonUtil.getParam(methodCall, result, "inviteeList").toString().split(","));
+        List<String> inviteeList = new ArrayList<>(Arrays.asList(CommonUtil.getParam(methodCall, result, "inviteeList").toString().split(",")));
         String data = CommonUtil.getParam(methodCall, result, "data");
         Boolean onlineUserOnly = CommonUtil.getParam(methodCall, result, "onlineUserOnly");
         int timeout = CommonUtil.getParam(methodCall, result, "timeout");
@@ -499,6 +552,50 @@ public class TencentImPlugin implements FlutterPlugin, MethodCallHandler {
                 @Override
                 public void onSuccess(V2TIMMessage message) {
                     V2TIMManager.getMessageManager().getGroupHistoryMessageList(groupID, count, message, resultCallBack);
+                }
+            });
+        }
+    }
+
+    /**
+     * 获得消息记录
+     *
+     * @param methodCall 方法调用对象
+     * @param result     返回结果对象
+     */
+    private void getHistoryMessageList(MethodCall methodCall, final Result result) {
+        final String groupID = methodCall.argument("groupID");
+        final String userID = methodCall.argument("userID");
+        final int count = CommonUtil.getParam(methodCall, result, "count");
+        final int type = CommonUtil.getParam(methodCall, result, "type");
+        String lastMsgStr = methodCall.argument("lastMsg");
+
+        // 返回回调对象
+        final ValueCallBack<List<V2TIMMessage>> resultCallBack = new ValueCallBack<List<V2TIMMessage>>(result) {
+            @Override
+            public void onSuccess(List<V2TIMMessage> v2TIMMessages) {
+                List<CustomMessageEntity> resultData = new ArrayList<>(v2TIMMessages.size());
+                for (V2TIMMessage v2TIMMessage : v2TIMMessages) {
+                    resultData.add(new CustomMessageEntity(v2TIMMessage));
+                }
+                result.success(JsonUtil.toJSONString(resultData));
+            }
+        };
+
+        // 根据是否传递最后一条消息进行特殊处理
+        final V2TIMMessageListGetOption opt = new V2TIMMessageListGetOption();
+        opt.setUserID(userID);
+        opt.setGroupID(groupID);
+        opt.setGetType(type);
+        opt.setCount(count);
+        if (lastMsgStr == null) {
+            V2TIMManager.getMessageManager().getHistoryMessageList(opt, resultCallBack);
+        } else {
+            TencentImUtils.getMessageByFindMessageEntity(lastMsgStr, new ValueCallBack<V2TIMMessage>(result) {
+                @Override
+                public void onSuccess(V2TIMMessage message) {
+                    opt.setLastMsg(message);
+                    V2TIMManager.getMessageManager().getHistoryMessageList(opt, resultCallBack);
                 }
             });
         }