+++ lib/tencent_im_plugin.dart
@@ -27,6 +27,7 @@ import 'package:tencent_im_plugin/entity/signaling_info_entity.dart';
 import 'package:tencent_im_plugin/entity/user_entity.dart';
 import 'package:tencent_im_plugin/enums/friend_application_agree_type_enum.dart';
 import 'package:tencent_im_plugin/enums/friend_type_enum.dart';
+import 'package:tencent_im_plugin/enums/get_message_type_enum.dart';
 import 'package:tencent_im_plugin/enums/group_member_filter_enum.dart';
 import 'package:tencent_im_plugin/enums/group_member_role_enum.dart';
 import 'package:tencent_im_plugin/enums/group_receive_message_opt_enum.dart';
@@ -39,7 +40,8 @@ import 'package:tencent_im_plugin/entity/message_entity.dart';
 import 'package:tencent_im_plugin/enums/log_print_level.dart';
 
 class TencentImPlugin {
-  static const MethodChannel _channel = const MethodChannel('tencent_im_plugin');
+  static const MethodChannel _channel =
+      const MethodChannel('tencent_im_plugin');
 
   /// 监听器对象
   static TencentImPluginListener listener = TencentImPluginListener(_channel);
@@ -66,6 +68,14 @@ class TencentImPlugin {
   /// 反初始化SDK
   static unInitSDK() => _channel.invokeMethod('unInitSDK');
 
+  /// 获取SDK版本
+  /// [Return] SDK版本
+  static Future<String> getVersion() => _channel.invokeMethod('getVersion');
+
+  /// 获取服务器当前时间
+  /// [Return] 服务器时间，单位 s
+  static Future<int> getServerTime() => _channel.invokeMethod('getServerTime');
+
   /// 登录腾讯云IM
   /// [userID] 用户ID
   /// [userSig] 用户签名
@@ -73,7 +83,8 @@ class TencentImPlugin {
     @required String userID,
     @required String userSig,
   }) {
-    return _channel.invokeMethod('login', {"userID": userID, "userSig": userSig});
+    return _channel
+        .invokeMethod('login', {"userID": userID, "userSig": userSig});
   }
 
   /// 退出登录腾讯云IM
@@ -86,7 +97,8 @@ class TencentImPlugin {
 
   /// 获得用户登录状态
   /// [Return] 用户当前登录状态
-  static Future<LoginStatusEnum> getLoginStatus() async => LoginStatusTool.getByInt(await _channel.invokeMethod('getLoginStatus'));
+  static Future<LoginStatusEnum> getLoginStatus() async =>
+      LoginStatusTool.getByInt(await _channel.invokeMethod('getLoginStatus'));
 
   /// 获得当前登录用户
   /// [Return] 当前用户ID
@@ -183,7 +195,8 @@ class TencentImPlugin {
   static Future<SignalingInfoEntity> getSignalingInfo({
     @required FindMessageEntity message,
   }) async {
-    return SignalingInfoEntity.fromJson(jsonDecode(await _channel.invokeMethod('getSignalingInfo', {
+    return SignalingInfoEntity.fromJson(
+        jsonDecode(await _channel.invokeMethod('getSignalingInfo', {
       "message": jsonEncode(message),
     })));
   }
@@ -228,7 +241,8 @@ class TencentImPlugin {
         "localCustomInt": localCustomInt,
         "localCustomStr": localCustomStr,
         "priority": MessagePriorityTool.toInt(priority),
-        "offlinePushInfo": offlinePushInfo == null ? null : jsonEncode(offlinePushInfo),
+        "offlinePushInfo":
+            offlinePushInfo == null ? null : jsonEncode(offlinePushInfo),
       }..removeWhere((key, value) => value == null),
     );
   }
@@ -265,7 +279,8 @@ class TencentImPlugin {
         "localCustomInt": localCustomInt,
         "localCustomStr": localCustomStr,
         "priority": MessagePriorityTool.toInt(priority),
-        "offlinePushInfo": offlinePushInfo == null ? null : jsonEncode(offlinePushInfo),
+        "offlinePushInfo":
+            offlinePushInfo == null ? null : jsonEncode(offlinePushInfo),
       }..removeWhere((key, value) => value == null),
     );
   }
@@ -280,23 +295,32 @@ class TencentImPlugin {
     });
   }
 
-  /// 获得历史记录，此方法为 getC2CHistoryMessageList + getGroupHistoryMessageList 提供的统一封装方法，(userID 和 group) 不能都为空
-  /// [userId] 用户ID
+  /// 获取历史消息高级接口
+  /// [userID] 用户ID
   /// [groupID] 群聊ID
   /// [count] 拉取消息的个数，不宜太多，会影响消息拉取的速度，这里建议一次拉取 20 个
   /// [lastMsg] 获取消息的起始消息，如果传 null，起始消息为会话的最新消息
+  /// [type] 拉取消息类型，可以设置拉取本地、云端更老或则更新的消息。
   static Future<List<MessageEntity>> getHistoryMessageList({
     String userID,
     String groupID,
     @required int count,
+    GetMessageTypeEnum type: GetMessageTypeEnum.GetCloudOlderMsg,
     FindMessageEntity lastMsg,
-  }) {
-    if (userID == null && groupID == null) throw ArgumentError("userID 和 groupID 不能同时为空!");
-    if (userID != null) {
-      return getC2CHistoryMessageList(userID: userID, count: count, lastMsg: lastMsg);
-    } else {
-      return getGroupHistoryMessageList(groupID: groupID, count: count, lastMsg: lastMsg);
-    }
+  }) async {
+    if (userID == null && groupID == null)
+      throw ArgumentError("userID 和 groupID 不能同时为空!");
+    return ListUtil.generateOBJList<MessageEntity>(
+        jsonDecode(await _channel.invokeMethod(
+      'getHistoryMessageList',
+      {
+        "userID": userID,
+        "groupID": groupID,
+        "count": count,
+        "type": GetMessageTypeTool.toInt(type),
+        "lastMsg": lastMsg == null ? null : jsonEncode(lastMsg),
+      }..removeWhere((key, value) => value == null),
+    )));
   }
 
   /// 获得单聊历史记录
@@ -308,7 +332,8 @@ class TencentImPlugin {
     @required int count,
     FindMessageEntity lastMsg,
   }) async {
-    return ListUtil.generateOBJList<MessageEntity>(jsonDecode(await _channel.invokeMethod(
+    return ListUtil.generateOBJList<MessageEntity>(
+        jsonDecode(await _channel.invokeMethod(
       'getC2CHistoryMessageList',
       {
         "userID": userID,
@@ -327,7 +352,8 @@ class TencentImPlugin {
     @required int count,
     FindMessageEntity lastMsg,
   }) async {
-    return ListUtil.generateOBJList<MessageEntity>(jsonDecode(await _channel.invokeMethod(
+    return ListUtil.generateOBJList<MessageEntity>(
+        jsonDecode(await _channel.invokeMethod(
       'getGroupHistoryMessageList',
       {
         "groupID": groupID,
@@ -344,7 +370,8 @@ class TencentImPlugin {
     String userID,
     String groupID,
   }) {
-    if (userID == null && groupID == null) throw ArgumentError("userID 和 groupID 不能同时为空!");
+    if (userID == null && groupID == null)
+      throw ArgumentError("userID 和 groupID 不能同时为空!");
     if (userID != null) {
       return markC2CMessageAsRead(userID: userID);
     } else {
@@ -496,7 +523,8 @@ class TencentImPlugin {
   static Future<List<MessageEntity>> findMessages({
     @required List<FindMessageEntity> messages,
   }) async {
-    return ListUtil.generateOBJList<MessageEntity>(await _channel.invokeMethod('findMessages', {
+    return ListUtil.generateOBJList<MessageEntity>(
+        await _channel.invokeMethod('findMessages', {
       "messages": jsonEncode(messages),
     }));
   }
@@ -553,7 +581,8 @@ class TencentImPlugin {
 
   /// 获取已经加入的群列表（不包括已加入的直播群）
   static Future<List<GroupInfoEntity>> getJoinedGroupList() async {
-    return ListUtil.generateOBJList<GroupInfoEntity>(jsonDecode(await _channel.invokeMethod('getJoinedGroupList')));
+    return ListUtil.generateOBJList<GroupInfoEntity>(
+        jsonDecode(await _channel.invokeMethod('getJoinedGroupList')));
   }
 
   /// 拉取群资料
@@ -561,7 +590,8 @@ class TencentImPlugin {
   static Future<List<GroupInfoResultEntity>> getGroupsInfo({
     @required List<String> groupIDList,
   }) async {
-    return ListUtil.generateOBJList<GroupInfoResultEntity>(jsonDecode(await _channel.invokeMethod('getGroupsInfo', {
+    return ListUtil.generateOBJList<GroupInfoResultEntity>(
+        jsonDecode(await _channel.invokeMethod('getGroupsInfo', {
       "groupIDList": groupIDList.join(","),
     })));
   }
@@ -674,7 +704,8 @@ class TencentImPlugin {
     GroupMemberFilterEnum filter: GroupMemberFilterEnum.All,
     int nextSeq: 0,
   }) async {
-    return GroupMemberInfoResultEntity.fromJson(jsonDecode(await _channel.invokeMethod('getGroupMemberList', {
+    return GroupMemberInfoResultEntity.fromJson(
+        jsonDecode(await _channel.invokeMethod('getGroupMemberList', {
       "groupID": groupID,
       "filter": GroupMemberFilterTool.toInt(filter),
       "nextSeq": nextSeq,
@@ -688,7 +719,8 @@ class TencentImPlugin {
     @required String groupID,
     @required List<String> memberList,
   }) async {
-    return ListUtil.generateOBJList<GroupMemberEntity>(jsonDecode(await _channel.invokeMethod('getGroupMembersInfo', {
+    return ListUtil.generateOBJList<GroupMemberEntity>(
+        jsonDecode(await _channel.invokeMethod('getGroupMembersInfo', {
       "groupID": groupID,
       "memberList": memberList.join(","),
     })));
@@ -730,7 +762,8 @@ class TencentImPlugin {
     @required String groupID,
     @required List<String> userList,
   }) async {
-    return ListUtil.generateOBJList<GroupMemberOperationResultEntity>(jsonDecode(await _channel.invokeMethod('inviteUserToGroup', {
+    return ListUtil.generateOBJList<GroupMemberOperationResultEntity>(
+        jsonDecode(await _channel.invokeMethod('inviteUserToGroup', {
       "groupID": groupID,
       "userList": userList.join(","),
     })));
@@ -745,7 +778,8 @@ class TencentImPlugin {
     @required List<String> memberList,
     String reason: "",
   }) async {
-    return ListUtil.generateOBJList<GroupMemberOperationResultEntity>(jsonDecode(await _channel.invokeMethod('kickGroupMember', {
+    return ListUtil.generateOBJList<GroupMemberOperationResultEntity>(
+        jsonDecode(await _channel.invokeMethod('kickGroupMember', {
       "groupID": groupID,
       "memberList": memberList.join(","),
       "reason": reason,
@@ -783,7 +817,8 @@ class TencentImPlugin {
 
   /// 获取加群的申请列表
   static Future<GroupApplicationResultEntity> getGroupApplicationList() async {
-    return GroupApplicationResultEntity.fromJson(jsonDecode(await _channel.invokeMethod('getGroupApplicationList')));
+    return GroupApplicationResultEntity.fromJson(
+        jsonDecode(await _channel.invokeMethod('getGroupApplicationList')));
   }
 
   /// 同意某一条加群申请
@@ -824,7 +859,8 @@ class TencentImPlugin {
     int nextSeq: 0,
     int count: 100,
   }) async {
-    return ConversationResultEntity.fromJson(jsonDecode(await _channel.invokeMethod('getConversationList', {
+    return ConversationResultEntity.fromJson(
+        jsonDecode(await _channel.invokeMethod('getConversationList', {
       "nextSeq": nextSeq,
       "count": count,
     })));
@@ -845,7 +881,8 @@ class TencentImPlugin {
     } else if (cID == null && groupID != null) {
       cID = conversationGroupPrefix + groupID;
     }
-    return ConversationEntity.fromJson(jsonDecode(await _channel.invokeMethod('getConversation', {
+    return ConversationEntity.fromJson(
+        jsonDecode(await _channel.invokeMethod('getConversation', {
       "conversationID": cID,
     })));
   }
@@ -901,7 +938,8 @@ class TencentImPlugin {
   static Future<List<UserEntity>> getUsersInfo({
     @required List<String> userIDList,
   }) async {
-    return ListUtil.generateOBJList<UserEntity>(jsonDecode(await _channel.invokeMethod('getUsersInfo', {
+    return ListUtil.generateOBJList<UserEntity>(
+        jsonDecode(await _channel.invokeMethod('getUsersInfo', {
       "userIDList": userIDList.join(","),
     })));
   }
@@ -921,7 +959,8 @@ class TencentImPlugin {
   static Future<List<FriendOperationResultEntity>> addToBlackList({
     @required List<String> userIDList,
   }) async {
-    return ListUtil.generateOBJList<FriendOperationResultEntity>(jsonDecode(await _channel.invokeMethod('addToBlackList', {
+    return ListUtil.generateOBJList<FriendOperationResultEntity>(
+        jsonDecode(await _channel.invokeMethod('addToBlackList', {
       "userIDList": userIDList.join(","),
     })));
   }
@@ -931,14 +970,16 @@ class TencentImPlugin {
   static Future<List<FriendOperationResultEntity>> deleteFromBlackList({
     @required List<String> userIDList,
   }) async {
-    return ListUtil.generateOBJList<FriendOperationResultEntity>(jsonDecode(await _channel.invokeMethod('deleteFromBlackList', {
+    return ListUtil.generateOBJList<FriendOperationResultEntity>(
+        jsonDecode(await _channel.invokeMethod('deleteFromBlackList', {
       "userIDList": userIDList.join(","),
     })));
   }
 
   /// 获得黑名单列表
   static Future<List<FriendInfoEntity>> getBlackList() async {
-    return ListUtil.generateOBJList<FriendInfoEntity>(jsonDecode(await _channel.invokeMethod('getBlackList')));
+    return ListUtil.generateOBJList<FriendInfoEntity>(
+        jsonDecode(await _channel.invokeMethod('getBlackList')));
   }
 
   /// 设置离线推送Token,Android使用setOfflinePushConfig，IOS使用setAPNS
@@ -966,7 +1007,8 @@ class TencentImPlugin {
 
   /// 获得好友列表
   static Future<List<FriendInfoEntity>> getFriendList() async {
-    return ListUtil.generateOBJList<FriendInfoEntity>(jsonDecode(await _channel.invokeMethod('getFriendList')));
+    return ListUtil.generateOBJList<FriendInfoEntity>(
+        jsonDecode(await _channel.invokeMethod('getFriendList')));
   }
 
   /// 获得指定好友信息
@@ -974,7 +1016,8 @@ class TencentImPlugin {
   static Future<List<FriendInfoResultEntity>> getFriendsInfo({
     @required List<String> userIDList,
   }) async {
-    return ListUtil.generateOBJList<FriendInfoResultEntity>(jsonDecode(await _channel.invokeMethod('getFriendsInfo', {
+    return ListUtil.generateOBJList<FriendInfoResultEntity>(
+        jsonDecode(await _channel.invokeMethod('getFriendsInfo', {
       "userIDList": userIDList.join(","),
     })));
   }
@@ -994,7 +1037,8 @@ class TencentImPlugin {
   static Future<FriendOperationResultEntity> addFriend({
     @required FriendAddApplicationEntity info,
   }) async {
-    return FriendOperationResultEntity.fromJson(jsonDecode(await _channel.invokeMethod('addFriend', {
+    return FriendOperationResultEntity.fromJson(
+        jsonDecode(await _channel.invokeMethod('addFriend', {
       "info": jsonEncode(info),
     })));
   }
@@ -1006,7 +1050,8 @@ class TencentImPlugin {
     @required List<String> userIDList,
     @required FriendTypeEnum deleteType,
   }) async {
-    return ListUtil.generateOBJList<FriendOperationResultEntity>(jsonDecode(await _channel.invokeMethod('deleteFromFriendList', {
+    return ListUtil.generateOBJList<FriendOperationResultEntity>(
+        jsonDecode(await _channel.invokeMethod('deleteFromFriendList', {
       "userIDList": userIDList.join(","),
       "deleteType": FriendTypeTool.toInt(deleteType),
     })));
@@ -1019,15 +1064,18 @@ class TencentImPlugin {
     @required String userID,
     @required FriendTypeEnum checkType,
   }) async {
-    return FriendCheckResultEntity.fromJson(jsonDecode(await _channel.invokeMethod('checkFriend', {
+    return FriendCheckResultEntity.fromJson(
+        jsonDecode(await _channel.invokeMethod('checkFriend', {
       "userID": userID,
       "checkType": FriendTypeTool.toInt(checkType),
     })));
   }
 
   /// 获取好友申请列表
-  static Future<FriendApplicationResultEntity> getFriendApplicationList() async {
-    return FriendApplicationResultEntity.fromJson(jsonDecode(await _channel.invokeMethod('getFriendApplicationList')));
+  static Future<FriendApplicationResultEntity>
+      getFriendApplicationList() async {
+    return FriendApplicationResultEntity.fromJson(
+        jsonDecode(await _channel.invokeMethod('getFriendApplicationList')));
   }
 
   /// 同意好友申请
@@ -1037,7 +1085,8 @@ class TencentImPlugin {
     @required FindFriendApplicationEntity application,
     @required FriendApplicationAgreeTypeEnum responseType,
   }) async {
-    return FriendOperationResultEntity.fromJson(jsonDecode(await _channel.invokeMethod('acceptFriendApplication', {
+    return FriendOperationResultEntity.fromJson(
+        jsonDecode(await _channel.invokeMethod('acceptFriendApplication', {
       "application": jsonEncode(application),
       "responseType": FriendApplicationAgreeTypeTool.toInt(responseType),
     })));
@@ -1048,7 +1097,8 @@ class TencentImPlugin {
   static Future<FriendOperationResultEntity> refuseFriendApplication({
     @required FindFriendApplicationEntity application,
   }) async {
-    return FriendOperationResultEntity.fromJson(jsonDecode(await _channel.invokeMethod('refuseFriendApplication', {
+    return FriendOperationResultEntity.fromJson(
+        jsonDecode(await _channel.invokeMethod('refuseFriendApplication', {
       "application": jsonEncode(application),
     })));
   }
@@ -1075,7 +1125,8 @@ class TencentImPlugin {
     @required String groupName,
     @required List<String> userIDList,
   }) async {
-    return ListUtil.generateOBJList<FriendOperationResultEntity>(jsonDecode(await _channel.invokeMethod('createFriendGroup', {
+    return ListUtil.generateOBJList<FriendOperationResultEntity>(
+        jsonDecode(await _channel.invokeMethod('createFriendGroup', {
       "groupName": groupName,
       "userIDList": userIDList.join(","),
     })));
@@ -1086,7 +1137,8 @@ class TencentImPlugin {
   static Future<List<FriendGroupEntity>> getFriendGroups({
     List<String> groupNameList,
   }) async {
-    return ListUtil.generateOBJList<FriendGroupEntity>(jsonDecode(await _channel.invokeMethod(
+    return ListUtil.generateOBJList<FriendGroupEntity>(
+        jsonDecode(await _channel.invokeMethod(
       'getFriendGroups',
       {
         "groupNameList": groupNameList?.join(","),
@@ -1124,7 +1176,8 @@ class TencentImPlugin {
     @required String groupName,
     @required List<String> userIDList,
   }) async {
-    return ListUtil.generateOBJList<FriendOperationResultEntity>(jsonDecode(await _channel.invokeMethod('addFriendsToFriendGroup', {
+    return ListUtil.generateOBJList<FriendOperationResultEntity>(
+        jsonDecode(await _channel.invokeMethod('addFriendsToFriendGroup', {
       "groupName": groupName,
       "userIDList": userIDList.join(","),
     })));
@@ -1133,11 +1186,13 @@ class TencentImPlugin {
   /// 从分组中删除好友
   /// [groupName] 组名
   /// [userIDList] 好友ID
-  static Future<List<FriendOperationResultEntity>> deleteFriendsFromFriendGroup({
+  static Future<List<FriendOperationResultEntity>>
+      deleteFriendsFromFriendGroup({
     @required String groupName,
     @required List<String> userIDList,
   }) async {
-    return ListUtil.generateOBJList<FriendOperationResultEntity>(jsonDecode(await _channel.invokeMethod('deleteFriendsFromFriendGroup', {
+    return ListUtil.generateOBJList<FriendOperationResultEntity>(
+        jsonDecode(await _channel.invokeMethod('deleteFriendsFromFriendGroup', {
       "groupName": groupName,
       "userIDList": userIDList.join(","),
     })));
@@ -1147,5 +1202,6 @@ class TencentImPlugin {
   static addListener(TencentImListenerValue func) => listener.addListener(func);
 
   /// 移除消息监听
-  static removeListener(TencentImListenerValue func) => listener.removeListener(func);
+  static removeListener(TencentImListenerValue func) =>
+      listener.removeListener(func);
 }